requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
jinja2==3.1.3
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.1
msal==1.26.0